+++ Setup/Patch/Data/AddPersonalisedOptionsAttribute.php
@@ -1,0 +1,93 @@
+<?php
+/**
+ * Copyright © MageMe. All rights reserved.
+ * See LICENSE for license terms, or https://mageme.com/license.
+ */
+declare(strict_types=1);
+
+namespace MageMe\EUWithdrawal\Setup\Patch\Data;
+
+use Magento\Catalog\Model\Product;
+use Magento\Eav\Model\Entity\Attribute\Backend\ArrayBackend;
+use Magento\Eav\Model\Entity\Attribute\ScopedAttributeInterface;
+use Magento\Eav\Setup\EavSetup;
+use Magento\Eav\Setup\EavSetupFactory;
+use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Framework\Setup\Patch\DataPatchInterface;
+
+class AddPersonalisedOptionsAttribute implements DataPatchInterface
+{
+    public const ATTRIBUTE = 'custom_made_options';
+
+    private const GROUP = 'EU Withdrawal Compliance';
+
+    /**
+     * Constructor.
+     *
+     * @param ModuleDataSetupInterface $moduleDataSetup
+     * @param EavSetupFactory $eavSetupFactory
+     */
+    public function __construct(
+        private readonly ModuleDataSetupInterface $moduleDataSetup,
+        private readonly EavSetupFactory $eavSetupFactory,
+    ) {
+    }
+
+    /**
+     * Get dependencies.
+     *
+     * @return array
+     */
+    public static function getDependencies(): array
+    {
+        return [AddWithdrawalProductAttributes::class];
+    }
+
+    /**
+     * Get aliases.
+     *
+     * @return array
+     */
+    public function getAliases(): array
+    {
+        return [];
+    }
+
+    /**
+     * Apply.
+     *
+     * @return self
+     */
+    public function apply(): self
+    {
+        /** @var EavSetup $eavSetup */
+        $eavSetup = $this->eavSetupFactory->create(['setup' => $this->moduleDataSetup]);
+        $eavSetup->addAttribute(
+            Product::ENTITY,
+            self::ATTRIBUTE,
+            [
+                'type' => 'text',
+                'backend' => ArrayBackend::class,
+                'frontend' => '',
+                'label' => 'Options That Make It Custom-Made (Art. 16(c))',
+                'input' => 'multiselect',
+                'class' => '',
+                'source' => '',
+                'global' => ScopedAttributeInterface::SCOPE_GLOBAL,
+                'visible' => true,
+                'required' => false,
+                'user_defined' => true,
+                'default' => '',
+                'searchable' => false,
+                'filterable' => false,
+                'comparable' => false,
+                'visible_on_front' => false,
+                'used_in_product_listing' => false,
+                'unique' => false,
+                'group' => self::GROUP,
+            ],
+        );
+
+        return $this;
+    }
+}